+++ bases/niagads/open_access_filer_api/core.py
@@ -1,7 +1,8 @@
 from niagads.open_access_api_common.app import AppFactory
 from niagads.open_access_filer_api.documentation import OPEN_API_SPEC
-from niagads.open_access_filer_api.routes.root import router as InfoRouter
+from niagads.open_access_filer_api.routes.root import router as RootRouter
 from niagads.open_access_filer_api.routes.track import router as TrackRouter
+from niagads.open_access_filer_api.routes.search import router as SearchRouter
 from niagads.open_access_filer_api.routes.service import router as ServiceRouter
 from niagads.settings.core import get_service_environment
 import uvicorn
@@ -12,8 +13,9 @@ appFactory = AppFactory(
 )
 
 # add the child routes
-appFactory.add_router(InfoRouter)
+appFactory.add_router(RootRouter)
 appFactory.add_router(TrackRouter)
+appFactory.add_router(SearchRouter)
 appFactory.add_router(ServiceRouter)
 """
 app.include_router(TrackRouter)
