+++ to-be-migrated/api/routes/genomics/models/genomics_track.py
@@ -136,8 +136,8 @@ class GenomicsTrackSummaryResponse(PagedResponseModel):
 
     def to_text(self, format: ResponseView, **kwargs):
         fields = (
-            self.response[0].get_field_names()
-            if len(self.response) > 0
+            self.data[0].get_field_names()
+            if len(self.data) > 0
             else GenericTrackSummary.get_model_fields()
         )
         return super().to_text(format, fields=fields)
